Adding one more node in front of previously existing EqualsNode
